+++ instrumentation/opentelemetry-instrumentation-dbapi/src/opentelemetry/instrumentation/dbapi/__init__.py
@@ -43,14 +43,13 @@ import typing
 
 import wrapt
 
-from opentelemetry import context as opentelemetry_context
 from opentelemetry import trace as trace_api
 from opentelemetry.instrumentation.dbapi.version import __version__
 from opentelemetry.instrumentation.utils import (
-    _add_sql_comment,
     _get_opentelemetry_values,
     unwrap,
 )
+from opentelemetry.instrumentation.sqlcommenter_utils import _add_sql_comment
 from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.trace import SpanKind, TracerProvider, get_tracer
 
@@ -430,18 +429,6 @@ class CursorTracer:
                         "opentelemetry_values", True
                     ):
                         commenter_data.update(**_get_opentelemetry_values())
-
-                    # Add flask related tags
-                    sqlcommenter_flask_values = (
-                        opentelemetry_context.get_value(
-                            "SQLCOMMENTER_FLASK_VALUES"
-                        )
-                        if opentelemetry_context.get_value(
-                            "SQLCOMMENTER_FLASK_VALUES"
-                        )
-                        else {}
-                    )
-                    commenter_data.update(**sqlcommenter_flask_values)
 
                     # Filter down to just the requested attributes.
                     commenter_data = {
