+++ instrumentation/opentelemetry-instrumentation-flask/tests/base_test.py
@@ -30,7 +30,7 @@ class InstrumentationTest:
     def _sqlcommenter_endpoint():
         current_context = context.get_current()
         sqlcommenter_flask_values = current_context.get(
-            "SQLCOMMENTER_FLASK_VALUES", {}
+            "SQLCOMMENTER_ORM_TAGS_AND_VALUES", {}
         )
         return sqlcommenter_flask_values
