+++ instrumentation/opentelemetry-instrumentation-sqlalchemy/tests/test_sqlcommenter.py
@@ -68,7 +68,7 @@ class TestSqlalchemyInstrumentationWithSQLCommenter(TestBase):
 
         current_context = context.get_current()
         sqlcommenter_context = context.set_value(
-            "SQLCOMMENTER_FLASK_VALUES", {"flask": 1}, current_context
+            "SQLCOMMENTER_ORM_TAGS_AND_VALUES", {"flask": 1}, current_context
         )
         context.attach(sqlcommenter_context)
